Exercise Complete
The DOM manipulation was pretty easy. What I really struggled with was the logic for pressing a button to start the game.
Something just wasn't connecting for me to press the button and trigger other functions.
This might have been because I found it difficult to use the debugger with an event listener.